+++ domains/certificates/Query.API/Program.cs
