+++ domains/certificates/Query.API/appsettings.Development.json
