+++ domains/certificates/Query.API/appsettings.json
